Move webhook_topic_validator into /lib
